Intake testing opmode

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Firmware/Systems/Intake.java
@@ -15,7 +15,7 @@ public class Intake {
         intake.setMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
     }
 
-    public void setIntakePower (int power){
+    public void setIntakePower (double power){
         intake.setPower(power);
     }
 }
